Added docker compose for development

+++ spec/dummy/db/schema.rb
@@ -25,13 +25,13 @@ ActiveRecord::Schema.define(version: 20170312214504) do
   end
 
   create_table "cavy_groups_policies", id: false, force: :cascade do |t|
-    t.integer "cavy_group_id", null: false
-    t.integer "cavy_policy_id", null: false
+    t.bigint "cavy_group_id", null: false
+    t.bigint "cavy_policy_id", null: false
   end
 
   create_table "cavy_groups_users", id: false, force: :cascade do |t|
-    t.integer "cavy_user_id", null: false
-    t.integer "cavy_group_id", null: false
+    t.bigint "cavy_user_id", null: false
+    t.bigint "cavy_group_id", null: false
   end
 
   create_table "cavy_item_groups", id: :serial, force: :cascade do |t|
